requirements.txt
# Hades Prediction Market - Dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pytz>=2023.0
certifi>=2023.0
PyYAML>=6.0
pytest>=7.0
pandas>=1.5.0
websockets>=12.0
